Revert "WIP refactor metadata service to use repository implementation"
This reverts commit 526b88f206e2cd85f7cdc88846d5b6d8eec9cfb3.
